+++ src/components/button/index.ts
@@ -2,5 +2,13 @@ import BackButton from "@/components/button/BackButton";
 import CheckButton from "@/components/button/CheckButton";
 import HomeButton from "@/components/button/HomeButton";
 import { PlusButton, MinusButton } from "@/components/button/QuantityButton";
+import NavigateButton from "./NavigateButton";
 
-export { BackButton, CheckButton, HomeButton, PlusButton, MinusButton };
+export {
+  BackButton,
+  CheckButton,
+  HomeButton,
+  PlusButton,
+  MinusButton,
+  NavigateButton,
+};
