--- src/components/section/Footer/styled.ts
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Container = styled.div`
-  height: 64px;
-  font-weight: var(--font-weight-bold);
-  line-height: 16px;
-`;
